refactor: create reusable email layout
