+++ lib/app.ts
@@ -3,6 +3,7 @@ import packageInfo from '../package.json';
 const app = {
   version: packageInfo.version,
   name: 'BoxyHQ',
+  location: 'Imaginary Location, London, UK',
   logoUrl: 'https://boxyhq.com/img/logo.png',
   url: 'http://localhost:4002',
 };
